[Decomp][Player] Extract SpellCooldownMgr from Player.cpp
